+++ core/settings.py
@@ -49,7 +49,9 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'line_bot'
+    'line_bot',
+    'users',
+    'cafe'
 ]
 
 MIDDLEWARE = [
